Extract non-AR concern
